requirements.txt
fastapi
uvicorn[standard]
python-dotenv
requests
websockets
openai
pyttsx3
pydub
